modules added index update on deletion fixed
