+++ src/modules/todo.js
@@ -1,0 +1,7 @@
+export default class toDoTask {
+  constructor(desc, index, completed = false) {
+    this.desc = desc;
+    this.index = index;
+    this.completed = completed;
+  }
+}
